+++ src/lib.rs
@@ -197,6 +197,7 @@ pub mod app_kit;
 pub mod core_animation;
 pub mod core_foundation;
 pub mod core_graphics;
+pub mod core_image;
 pub mod dispatch;
 pub mod foundation;
 pub mod ui_kit;